backend/requirements.txt
# FastAPI Backend Dependencies
# ============================

# Web Framework
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.6

# Configuration
python-dotenv>=1.2.0
PyYAML>=6.0.3

# Database
neo4j>=6.1.0

# AI/LLM
google-genai>=1.59.0
openai>=1.30.0
anthropic>=0.30.0

# Data Validation
pydantic>=2.12.0

# HTTP Client
httpx>=0.28.0

# Authentication
python-jose[cryptography]>=3.3.0
